Update cred and source tests to test other types
While working on stubbing out test cases for vcenter and satellite
sources, I realized I could stack @pytest.mark.parametrize and kill two
birds with one stone.

Currently no satellite functionality is implemented in quipucords, so I
introduced QCS_SOURCE_TYPES in camayoc.constants which can be updated at a later date to
include the 'satellite' type.

This commit updates credential and source tests where they could be
generalized, and stubs have been created or test moved to specific
modules if the tests only applied to a certain type.

Closes #91
