+++ camayoc/constants.py
@@ -323,3 +323,6 @@ QCS_SOURCE_PATH = 'sources/'
 
 QCS_SCAN_PATH = 'scans/'
 """The path to the scans endpoint for CRUD tasks."""
+
+QCS_SOURCE_TYPES = ['vcenter', 'network']
+"""Types of sources that the quipucords server supports."""
