+++ docs/api/camayoc.tests.qcs.api.credentials.rst
@@ -11,7 +11,6 @@ Submodules
 
 .. toctree::
 
+   camayoc.tests.qcs.api.credentials.test_creds_common
    camayoc.tests.qcs.api.credentials.test_network_creds
-   camayoc.tests.qcs.api.credentials.test_sat_creds
-   camayoc.tests.qcs.api.credentials.test_vcenter_creds
