+++ docs/api/camayoc.tests.qcs.api.credentials.test_creds_common.rst
@@ -1,0 +1,7 @@
+camayoc\.tests\.qcs\.api\.credentials\.test\_creds\_common module
+=================================================================
+
+.. automodule:: camayoc.tests.qcs.api.credentials.test_creds_common
+    :members:
+    :undoc-members:
+    :show-inheritance:
