--- docs/api/camayoc.tests.qcs.api.credentials.test_sat_creds.rst
@@ -1,7 +1,0 @@
-camayoc\.tests\.qcs\.api\.credentials\.test\_sat\_creds module
-==============================================================
-
-.. automodule:: camayoc.tests.qcs.api.credentials.test_sat_creds
-    :members:
-    :undoc-members:
-    :show-inheritance:
