--- docs/api/camayoc.tests.qcs.api.credentials.test_vcenter_creds.rst
@@ -1,7 +1,0 @@
-camayoc\.tests\.qcs\.api\.credentials\.test\_vcenter\_creds module
-==================================================================
-
-.. automodule:: camayoc.tests.qcs.api.credentials.test_vcenter_creds
-    :members:
-    :undoc-members:
-    :show-inheritance:
