+++ docs/api/camayoc.tests.qcs.api.sources.rst
@@ -12,4 +12,6 @@ Submodules
 .. toctree::
 
    camayoc.tests.qcs.api.sources.test_network_sources
+   camayoc.tests.qcs.api.sources.test_sources_common
+   camayoc.tests.qcs.api.sources.test_vcenter_sources
