+++ docs/api/camayoc.tests.qcs.api.sources.test_sources_common.rst
@@ -1,0 +1,7 @@
+camayoc\.tests\.qcs\.api\.sources\.test\_sources\_common module
+===============================================================
+
+.. automodule:: camayoc.tests.qcs.api.sources.test_sources_common
+    :members:
+    :undoc-members:
+    :show-inheritance:
