+++ docs/api/camayoc.tests.qcs.api.sources.test_vcenter_sources.rst
@@ -1,0 +1,7 @@
+camayoc\.tests\.qcs\.api\.sources\.test\_vcenter\_sources module
+================================================================
+
+.. automodule:: camayoc.tests.qcs.api.sources.test_vcenter_sources
+    :members:
+    :undoc-members:
+    :show-inheritance:
